Allow the army definition use a string for a base.
The string will later be used for retrieving extra information about the
base when the game is saved and loaded.  This is in contrast to having
the base name, which is displayed to the user, be used as the key
for determining the state of the base.  For example, mobile, General,
...