Handle tool tips for battle cards
A base that has a battle card must be declared as a variable,
and the variable name used as a string in the army.

The string is converted to the table when spawning the army,
but the variable name is retained in extra data that is used
to build up the tool tip after the army has already spawned.

By using the decorator design pattern, the extra data can
be persisted.  If the extra data was assiged to the object
using Object.setVar, the data would not be persisted.