Change history to a doubly linked list and snapshots
Instead of using a map a double linked list is used to store the
history.  The methods have been renamed to make it clearer
what is the purpose.

In addition the history now has a capacity so only 60 events
are remembered versus remembering the entire game.  This should help
the late game sluggishness that is seen.

Instead of storing the transformation of the bases, the position and
rotation is stored.  This allows us to be able to undo the group
moves by dragging.  Rotations are also able to be undone.

If a base has not moved since its last snapshot, it is not placed into
the snapshot, this should give us the same memory consumption per
history event has it currently being used, or possibly a little less.

As a place holder for future history events of a different type,
e.g. dismounting, a type attribute has been added.  The event
it self is responsible for restoring its state when we
undo/redo.