Refactor. Remove spawning static zones.
The zones for counting dead and army builder are in the
saved TTS file.  Stop deleting them and creating them
again.

The creation was used to load g_dead_zones, a map from zone
name to the zone.  g_dead_zones has been removed and instead
we get the zone by object name.  The only time it is used
is when we use the context object menu to kill a base.