Make the generation of the note for the army tolerant of missing data
Factors such as invasion are not yet handled by the meshwesh interface,
for armies that have variants, where the variants can have different
values.

In this case we return nil, and the note has to handle it.

The note will just leave a blank field when the value is nil.