Replace notecard with army details with tablet.
An army page on Meshwesh can be very long, for example
Later Hebrews stretches out to 8 pages when you include the
allies.

Rather than trying to fit that on an index card, or
as a PDF a tablet is used.  When an army is loaded the
Meshwesh page for that army is loaded.

As part of this change the legacy, V1, army definitions have
been removed.