History to handle delting a base
You can now undo a base being destructed which is used when a base
is dismounted.

Rewrote the history mechanism to record the old and new positions
of bases that are moved, instead of using a snapshot.

All history events now have to have undo() and redo().  Much more
like the original DBA mechanism. But we still support undoing
rotations and moves as a group, back to their original position.