Implement left and right turn to face.
Base does not fall back automatically, I wanted it to be
two actions, to make it obvious when a collision happens.

A turn to face is a 90 degree rotation, with the side edge
being on the line the front edge was before, and with
the front edge remaining in contact with the enemy.