Initial UI changes for deploying allies
Allies will be deployed on the screen as the main army.
This will allow a common filtering by date, and recording of the
main army so we can filter choices in later wizard steps.

The first wizard step to select between Grand Triumph! and
Triumph! is part of this change.