Rename ZOC, Rulers, Snap text on DBA orders panel
Shrink the space needed for text ZOC, Rulers, and Snap
to not have the "Do not show" in front.

Now it is a much more condensed "ZOC", "Rulers", "Show Bases"
with no changing of the text.

Makes the movement slider bigger, but it should be shrunk a bit
to give padding.