Tatical Advantage Wizard
Use the maneuver rating from Meshwesh, with an option box in case
there is more than one.

Handle the die rolls.

Calculate the tatical advantage, handling ties.

First die roll is used to see if the army is eligible for a
flank march.  If so a Flank March card is placed in front of
the player.