Fix bug where wizard background frame was always present when loadng a game where the wizard had been completed.
As part of the fix the wizard state is now saved and
restored, so a saved game can have the wizard be
in the middle of a panel.

Another part of the fix is that save files that are
to be used for uploading to Steam are now cleaned
to remove the game state from them, so it is always as
if it is a new game.