Simplify defining of models by having common tile for army
An error previously occurred where the size of the tile
was not correct for the troop type.

Rather than specifying the tile size for every base
definition, the style is specified for the entire army,
e.g. base = "tile_grass".  The suffix, e.g. "_40x20" is then
calculated automatically based on the troop type.

The specific base can be overridden if desired in the model.