Remove large-table code
In DBA there was an option to have the table 33% larger, when
playing with larger bases.

The option was disabled for Triumph but the code became actived
and caused a bug that had the table too large.

Remove all the code around large-option so it cannot be activated
again, and simplifying the code base.