Reorganize dead using mat location.
Label the mat that the dead zone is on.
The mat rotates the right way for blue and red, opposite
each other, earlier they use to face the same direction.
Using the mat's location and rotation automatically
calculate how bases are placed on the mat when they are killed.
There is no longer any need to know if the mat is red or blue.

It should make it a little easier if we decide to move the mat.

Fixes bug where rotation of the mat, caused the dead pieces to
not be placed correctly.