Fix bug when more than 1 maneuver rating
Take an army that has more than one manuver rating such
as Viking Seas/Anglo-Danish.

When getting to the Tactical Advantage page before
rolling change the rating.

In the chat window an error use to be shown.

Roll, go to the next page, and come back again.
The rating was lost in the UI.