Wizard save uses enter method
There was a bug where if a game was saved with a wizard
panel that had state information, for example "Invasion"
which registers to listen to dice, when the game was loaded
the wizard panel was being showed by the state was lost.
In the case of "Invasion" dice rolls were not being recorded.

The change to the way panels are defined uses an entry and
exit function.  The exit function registration is now
static instead of dynamic.  The entry function is called
when the panel is being shown, including from onLoad.