Trees in Woods auto hide
When a base enters woods, all the trees are removed to make
it easy to see the base.  When the base is no longer in the
woods the trees come back.

Simmilar to how you would move trees off the table to get
them out of the way.

Also works for wooded hills.