Use github for assets
If the source code is on github at the current verion,
then use github to serve the assets.

The URL on github includes the change id, so if the
source code or asset is updated the URL is changed.

It makes it easier to modify assets since you do
not have to do a manual step of uploading the assets.