Do not include terrain point on border of area
If we are close to the border of a terrain area,
do not include a point.

Example: rocks for marsh should not spawn right next to the
edge of rough.

Fixes where a corner of the box checking if a point is
an area would just cross over the edge of the terrain area,
the point would be included.  This would cause terrain features
to pop up just outside the terrain.