Prepared defenses.
Plain bases spawned if army has battle card.  Counts
in army builder points.  Harmless if left in victory points.
Tool tip.

No 3d models.  Do not handle special cases where prepared defenses
are only for a special troop or campaign, but it is in the tool
tip.

The model for the prepared defenses is extremely thing, like
terrain, so troops can go onto it.