Reduce the loose troop model rotation
Loose troops models are placed at random facings
on a base to show that they are open ordered.
However when the rotation is too much the facing of the
base is hard to determine.

Reduce the random facing range from old range of -70 to +70
to new range of -50 to +50 degrees.