Fix base stuck on hill after reset snap
When a base is not moved a sufficient distance it is snapped
back to its initial position if snapping is enabled.

The base could be snapped such that it would not be able to
be picked up again, with the base stuck on the hill.
To workaround it in the game you could kill the base and bring it back.

The problem is solved by storing the initial base rotation when the
base is picked up and using that rotation when the base is reset.
The legacy solution was storing only the y-axis rotation in radians.