Fix bug where history step was lost
If you hit the move button twice, there was
only one undo and both moves got undone at
once.  You were only able to redo once, and
you lost the second move.

Move the recording of the history to the end
of the functions, so the history always
recorded the changes in the function.  Bug was
that the recording could take place too early since
the call to actually form the the history stack entry
was asynchronous.