Reset move distance on first base being selected.
If selecting a new set of bases, change the distance
to move to 0.

Goal is to make it easier when checking for firing and zoc
distance after moves, you select and base and see what the
actual position is versus some possible move.

Tradeoff is that is it more painful to move many groups the
same distance.

https://github.com/ArkeinGuardia/Triumph_TTS/issues/44
Shooting template arc of fire display suggestion #44