Kludge to redraw ZOC after rotating object.
Since the object is still in motion we cannot redraw
the ZOC on being notified that the object is about
to redraw.  So we turn off the gizmos, including the ZOC, and
then schedule a co-routine to redraw the gizmos in a little bit.

It causes the ZOCs to flash but at least they are correct.