Fix final bugs that allows zoc hiding in combat
When a base is in close combat then the base
has no ZOC. See Triumph! 51.4.b

Close combat is defined as having
and edge that is in contact with the front edge of
an enemy, or the front edge is in catact with
and enemy, and the bases are aligned.