Do not shrink ZOC when snapped side by size
When snapped the coordinates of the bases on the sides
are exactly the same, so it was dectected and the ZOC
shrunk.

As a kludge when we check if a base is blocking a zoc, we
shrink the base by a tiny bit so there are inequalities
in the pixels of the corners.