Suppress flank and rear ZOC for warwagon
If a friendly base is close, reduce the ZOC.
If in combat then there should be no ZOC.

Simplify the code, no need to keep track of who is being blocked
since we use distance to figure out when the zoc needs to be
recalculated.

Remove getting zoc for a base, not needed. Instead we
set the transform when a base moves.